Added reset color option, Added more terminals; Starts with, change to color

+++ src/main/java/mrfast/sbf/utils/Utils.java
@@ -173,7 +173,7 @@ public class Utils {
 		return 50D;
 	}
     
-    public static double randomNumber(int min,int max) {
+    public static int randomNumber(int min,int max) {
         Random rand = new Random();
         return rand.nextInt((max - min) + 1) + min;
     }
